backend/images/corona/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make images/corona/requirements.txt
#
attrs==19.3.0             # via pytest
azure-iot-device==2.1.2
azure-iot-hub==2.2.1
cachetools==4.1.0         # via google-auth
certifi==2020.4.5.2       # via msrest, requests, uamqp
cffi==1.14.0              # via cryptography
chardet==3.0.4            # via requests
coverage==5.1             # via pytest-cov
cryptography==2.9.2
google-api-core==1.19.1   # via opencensus
google-auth==1.16.1       # via google-api-core
googleapis-common-protos==1.52.0  # via google-api-core
graphviz==0.14            # via objgraph
idna==2.9                 # via requests
importlib-metadata==1.6.1  # via pluggy, pytest
isodate==0.6.0            # via msrest
janus==0.4.0              # via azure-iot-device
more-itertools==8.3.0     # via pytest
msrest==0.6.15            # via azure-iot-hub
oauthlib==3.1.0           # via requests-oauthlib
objgraph==3.4.1
opencensus-context==0.1.1  # via opencensus
opencensus-ext-azure==1.0.2
opencensus==0.7.7         # via opencensus-ext-azure
packaging==20.4           # via pytest
paho-mqtt==1.5.0          # via azure-iot-device
pluggy==0.13.1            # via pytest
prometheus-client==0.7.1  # via tornado-prometheus
protobuf==3.12.2          # via google-api-core, googleapis-common-protos
psutil==5.7.0             # via opencensus-ext-azure
py==1.8.1                 # via pytest
pyasn1-modules==0.2.8     # via google-auth
pyasn1==0.4.8             # via pyasn1-modules, rsa
pycparser==2.20           # via cffi
pyjwt==1.7.1
pyodbc==4.0.30
pyparsing==2.4.7          # via packaging
pysocks==1.7.1            # via azure-iot-device
pytest-asyncio==0.12.0
pytest-cov==2.9.0
pytest-tornado==0.8.0
pytest==5.4.3             # via pytest-asyncio, pytest-cov, pytest-tornado
python-dateutil==2.8.1
pytz==2020.1              # via google-api-core
redis==3.5.3
requests-oauthlib==1.3.0  # via msrest
requests-unixsocket==0.2.0  # via azure-iot-device
requests==2.23.0          # via azure-iot-device, google-api-core, msrest, opencensus-ext-azure, requests-oauthlib, requests-unixsocket
rsa==4.0                  # via google-auth
six==1.15.0               # via azure-iot-device, cryptography, google-api-core, google-auth, isodate, packaging, protobuf, python-dateutil, transitions, uamqp
testfixtures==6.14.1
tornado-prometheus==0.1.1
tornado==6.0.4
transitions==0.8.1        # via azure-iot-device
uamqp==1.2.8              # via azure-iot-hub
urllib3==1.25.9           # via azure-iot-device, requests, requests-unixsocket
wcwidth==0.2.4            # via pytest
zipp==3.1.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
